fix(types): fix SetAlphaParams.alpha type from Int32 to Double
M11: SetAlphaParams.alpha in NativeLayerBitmapTile.ts was declared as
Int32 but should be Double — the Java implementation calls
params.getDouble('alpha') and alpha is a float between 0 and 1.
Matches NativeLayerMBTilesBitmap.ts SetAlphaParams.alpha which
already uses Double.

+++ src/NativeModules/NativeLayerBitmapTile.ts
@@ -37,7 +37,7 @@ interface UpdateEnabledZoomMinMaxParams {
 interface SetAlphaParams {
 	nativeNodeHandle: Int32;
 	uuid: string;
-	alpha?: Int32;
+	alpha?: Double;
 }
 
 export type LayerBitmapTileProps = {
